Move 18ZOO to bundle
